feat: Add authentication controllers, services and implements security with jwt authentication based on OTP

+++ src/main/java/com/gauri/otpBasedAuthentication/dto/AuthResponse.java
@@ -18,7 +18,7 @@ public class AuthResponse {
     public static class AuthData {
         private String accessToken;
         private String refreshToken;
-        private Integer expiresIn;
+        private long expiresIn;
         private String tokenType;
         private UserResponse user;
         private Integer expiresIn_otp;
